plugin tests in separate files
